+++ Sistema/Smartgym/Domain/DTO/Exercicio.cs
@@ -12,14 +12,14 @@ namespace Domain.DTO
 
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int IdExercicio { get; set; }
+        public long IdExercicio { get; set; }
 
-        public int IdAparelhoExercicio { get; set; }
+        public long IdAparelhoExercicio { get; set; }
 
         [ForeignKey("IdAparelhoExercicio")]
         public virtual Exercicio AparelhoExercicio { get; set; }
 
-        public int IdSerieExercicio { get; set; }
+        public long IdSerieExercicio { get; set; }
 
         [ForeignKey("IdSerieExercicio")]
         public virtual Serie SerieExercicio { get; set; }
