+++ Sistema/Smartgym/Domain/DTO/Unidade.cs
@@ -13,9 +13,9 @@ namespace Domain.DTO
 
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int IdUnidade { get; set; }
+        public long IdUnidade { get; set; }
 
-        public int IdEnderecoUnidade { get; set; }
+        public long IdEnderecoUnidade { get; set; }
 
         [ForeignKey("IdEnderecoUnidade")]
         public virtual Endereco EnderecoUnidade { get; set; }
