+++ Sistema/Smartgym/Smartgym/Controllers/FichaController.cs
@@ -16,7 +16,7 @@ namespace Smartgym.Controllers
         }
 
         // GET: Ficha/Details/5
-        public ActionResult Details(int id)
+        public ActionResult Details(long id)
         {
             return View();
         }
@@ -45,7 +45,7 @@ namespace Smartgym.Controllers
         }
 
         // GET: Ficha/Edit/5
-        public ActionResult Edit(int id)
+        public ActionResult Edit(long id)
         {
             return View();
         }
@@ -53,7 +53,7 @@ namespace Smartgym.Controllers
         // POST: Ficha/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit(int id, IFormCollection collection)
+        public ActionResult Edit(long id, IFormCollection collection)
         {
             try
             {
@@ -68,7 +68,7 @@ namespace Smartgym.Controllers
         }
 
         // GET: Ficha/Delete/5
-        public ActionResult Delete(int id)
+        public ActionResult Delete(long id)
         {
             return View();
         }
@@ -76,7 +76,7 @@ namespace Smartgym.Controllers
         // POST: Ficha/Delete/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Delete(int id, IFormCollection collection)
+        public ActionResult Delete(long id, IFormCollection collection)
         {
             try
             {
